Fix handling of closed client's socket descriptors
Use pointer arithmetic instead of copying by value. We want to be
able to access the elements of the array (struct pollfd[]) passed to
poll. When a file descriptor is closed then we want to set the fd
field of pollfd instance to e.g. -1 (overwrite the actual file
descriptor that has been observed). Before this commit this has been
done by modifying an instance copied by value which was obviously a bug.